feat: seat id added

+++ booking-service/api/v1/routes/booking_seats.py
@@ -40,30 +40,41 @@ async def get_all_booking_seats(showing_id:str,db: Session = Depends(get_db)) ->
 
     showing_id_str = str(showing_id)      
     locked_seats = await redis_client.keys(f"locked_seat:{showing_id_str}:*")
-    locked_seats = [LockedSeatResponse(seat_id=UUID(ls.split(":")[2]), showing_id=UUID(ls.split(":")[1])) for ls in locked_seats]
+
+
+    locked_seats = [LockedSeatResponse(showing_id=UUID(ls.split(":")[1]),seat_id=UUID(ls.split(":")[2]), user_id=UUID(ls.split(":")[3])) for ls in locked_seats]
+
+    
     return ShowingSeatsResponse(booked_seats=[BookingSeatResponse.model_validate(bs) for bs in booking_seats],locked_seats=locked_seats)
 
 
 
 
 @router.post("/lock",status_code=status.HTTP_201_CREATED,summary="Lock a seat",response_model=BookingLockResponse)
-async def lock_seat(locked_seat: BookingLockCreate) -> BookingLockResponse:
+async def lock_seat(request: Request,locked_seat: BookingLockCreate) -> BookingLockResponse:
+    user_id = request.state.user_id
+    if user_id is None:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED,detail="Unauthorized")
     redis_client = await get_redis()
     if redis_client is None:
         raise HTTPException(status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,detail="Redis connection failed")
+    # Convert user_id string to UUID
+    user_id_uuid = UUID(user_id)
     # Convert UUIDs to strings for Redis (Redis requires string, bytes, int, or float)
     showing_id_str = str(locked_seat.showing_id)
     seat_id_str = str(locked_seat.seat_id)
+    user_id_str = str(user_id_uuid)
 
     #set the seat as locked for 10 minutes
     if locked_seat.lock_seat:
-        await redis_client.set(f"locked_seat:{showing_id_str}:{seat_id_str}", seat_id_str, ex=600)
+        await redis_client.set(f"locked_seat:{showing_id_str}:{seat_id_str}:{user_id_str}", seat_id_str, ex=600)
     else:
-        await redis_client.delete(f"locked_seat:{showing_id_str}:{seat_id_str}")
+        await redis_client.delete(f"locked_seat:{showing_id_str}:{seat_id_str}:{user_id_str}")
     return BookingLockResponse(
         # id=locked_seat.id,
         seat_id=locked_seat.seat_id,
-        showing_id=locked_seat.showing_id
+        showing_id=locked_seat.showing_id,
+        user_id=user_id_uuid
     )
